backend/requirements.txt
flask
mysql-connector-python
flask-bcrypt
python-dotenv
my